config: MODULES=dep and COMPRESS=lzma
COMPRESS=lzma compresses the initramfs better and Linux decompresses it
almost as fast as a gzip one would. It will fall back to gzip if the
initramfs can't be lzma compressed as long as config-`uname -r` exists
in /boot (or if that config cannot be found) so it is safe to enable
by default.

MODULES=dep may create 'unbootable' initramfs if you aren't booted
on a native system with the expected root device mounted with the
expected root device/filesystem modules loaded, but this isn't a
real problem and creates a much, much smaller initramfs.
Signed-off-by: Matt Sealey <matt@genesi-usa.com>